update docs and tests
